backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==1.10.13
python-dotenv==1.0.0
httpx==0.24.1
SQLAlchemy==1.4.53
psycopg2-binary==2.9.9
clickhouse-connect==0.6.23
orjson==3.9.10
loguru==0.7.2
python-multipart==0.0.6
ujson==5.8.0
tenacity==8.2.3
types-requests==2.31.0.10
pandas==1.5.3
lxml==4.9.3
openpyxl==3.1.2
alembic==1.12.1

